Refactored impulse
