+++ public/js/impulse.js
@@ -1,0 +1,108 @@
+const beatMod = require('./beat')
+const drums = require('./drummachine')
+
+
+var impulseResponseInfoList = [
+    // Impulse responses - each one represents a unique linear effect.
+    {"name":"No Effect", "url":"undefined", "dryMix":1, "wetMix":0},
+    {"name":"Spreader 2", "url":"impulse-responses/noise-spreader1.wav",        "dryMix":1, "wetMix":1},
+    {"name":"Spring Reverb", "url":"impulse-responses/feedback-spring.wav",     "dryMix":1, "wetMix":1},
+    {"name":"Space Oddity", "url":"impulse-responses/filter-rhythm3.wav",       "dryMix":1, "wetMix":0.7},
+    {"name":"Huge Reverse", "url":"impulse-responses/matrix6-backwards.wav",    "dryMix":0, "wetMix":0.7},
+    {"name":"Telephone Filter", "url":"impulse-responses/filter-telephone.wav", "dryMix":0, "wetMix":1.2},
+    {"name":"Lopass Filter", "url":"impulse-responses/filter-lopass160.wav",    "dryMix":0, "wetMix":0.5},
+    {"name":"Hipass Filter", "url":"impulse-responses/filter-hipass5000.wav",   "dryMix":0, "wetMix":4.0},
+    {"name":"Comb 1", "url":"impulse-responses/comb-saw1.wav",                  "dryMix":0, "wetMix":0.7},
+    {"name":"Comb 2", "url":"impulse-responses/comb-saw2.wav",                  "dryMix":0, "wetMix":1.0},
+    {"name":"Cosmic Ping", "url":"impulse-responses/cosmic-ping-long.wav",      "dryMix":0, "wetMix":0.9},
+    {"name":"Kitchen", "url":"impulse-responses/house-impulses/kitchen-true-stereo.wav", "dryMix":1, "wetMix":1},
+    {"name":"Living Room", "url":"impulse-responses/house-impulses/dining-living-true-stereo.wav", "dryMix":1, "wetMix":1},
+    {"name":"Living-Bedroom", "url":"impulse-responses/house-impulses/living-bedroom-leveled.wav", "dryMix":1, "wetMix":1},
+    {"name":"Dining-Far-Kitchen", "url":"impulse-responses/house-impulses/dining-far-kitchen.wav", "dryMix":1, "wetMix":1},
+    {"name":"Medium Hall 1", "url":"impulse-responses/matrix-reverb2.wav",      "dryMix":1, "wetMix":1},
+    {"name":"Medium Hall 2", "url":"impulse-responses/matrix-reverb3.wav",      "dryMix":1, "wetMix":1},
+    {"name":"Peculiar", "url":"impulse-responses/peculiar-backwards.wav",       "dryMix":1, "wetMix":1},
+    {"name":"Backslap", "url":"impulse-responses/backslap1.wav",                "dryMix":1, "wetMix":1},
+    {"name":"Diffusor", "url":"impulse-responses/diffusor3.wav",                "dryMix":1, "wetMix":1},
+    {"name":"Huge", "url":"impulse-responses/matrix-reverb6.wav",               "dryMix":1, "wetMix":0.7},
+]
+
+var impulseResponseList = 0;
+
+function setImpulseResponseList (lst) {
+  impulseResponseList = lst;
+  exports.impulseResponseList = impulseResponseList;
+}
+
+// function setImpulseResponseList_i (i, response) {
+//   impulseResponseList[i] =
+//   exports.impulseResponseList = impulseResponseList;
+// }
+
+
+
+function ImpulseResponse(url, index) {
+    this.url = url;
+    this.index = index;
+    this.startedLoading = false;
+    this.isLoaded_ = false;
+    this.buffer = 0;
+
+    this.demoIndex = -1; // no demo
+}
+
+ImpulseResponse.prototype.setDemoIndex = function(index) {
+    this.demoIndex = index;
+}
+
+ImpulseResponse.prototype.isLoaded = function() {
+    return this.isLoaded_;
+}
+
+function loadedImpulseResponse(buffer) {
+    this.buffer = buffer;
+    this.isLoaded_ = true;
+
+    if (this.demoIndex != -1) {
+        beatMod.beatDemo[this.demoIndex].setEffectLoaded();
+    }
+}
+
+ImpulseResponse.prototype.load = function() {
+    if (this.startedLoading) {
+        return;
+    }
+
+    this.startedLoading = true;
+
+    // Load asynchronously
+    var request = new XMLHttpRequest();
+    request.open("GET", this.url, true);
+    request.responseType = "arraybuffer";
+    this.request = request;
+
+    var asset = this;
+
+    request.onload = function() {
+        drums.context.decodeAudioData(request.response, loadedImpulseResponse.bind(asset) );
+    }
+
+    request.send();
+}
+
+
+
+// classes
+exports.ImpulseResponse = ImpulseResponse;
+
+// functions
+exports.setImpulseResponseList = setImpulseResponseList;
+
+// variables
+exports.impulseResponseList = impulseResponseList;
+exports.impulseResponseInfoList = impulseResponseInfoList;
+
+
+
+
+//
